deleted old CLIKommander and replaced it for CLICommand sources
